tests/py/requirements.txt
# This file is specifically to install correct version of libraries during CI testing.
# networkx library issue: https://discuss.pytorch.org/t/installing-pytorch-under-python-3-8-question-about-networkx-version/196740
expecttest
networkx
numpy
setuptools
parameterized>=0.2.0
pytest>=8.2.1
pytest-xdist>=3.6.1
pyyaml
transformers==4.53.1
nvidia-modelopt[all]; python_version >'3.9' and python_version <'3.13'
--extra-index-url https://pypi.nvidia.com
# flashinfer-python is not supported for python version 3.13 or higher
# flashinfer-python is broken on python 3.9 at the moment, so skip it for now
flashinfer-python; python_version >'3.9' and python_version <'3.13'
--extra-index-url https://download.pytorch.org/whl/nightly/cu130
torchvision>=0.26.0.dev,<0.27.0
timm>=1.0.3